+++ Changes.md
@@ -5,6 +5,7 @@ Version 3.3.0.2
   2. Fix too many anti-gravity icons.
   3. Fix unselected rows in the grading table.
   4. Fix button 'icon' colour.
+  5. Fix file picker icons.
 
 Version 3.3.0.1
   1. Beta for Moodle 3.3.