+++ Changes.md
@@ -1,5 +1,8 @@
 Version Information
 ===================
+Version 3.4.0.1
+  1. Beta for Moodle 3.4 with user tour reset and activity navigation.
+
 Version 3.3.0.2
   1. Fix user menu icons.
   2. Fix too many anti-gravity icons.